pe-org-air-platform/requirements.txt

# FastAPI & Server

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
starlette>=0.35.0
streamlit>=1.30.0
pytest
sec-edgar-downloader
sec-edgar-downloader
pdfplumber>=0.9.0
beautifulsoup4
lxml
pdfplumber
PyMuPDF

# Pydantic & Settings

pydantic>=2.5.0
pydantic-settings>=2.1.0


# Snowflake & Database

snowflake-connector-python>=3.6.0
snowflake-sqlalchemy>=1.5.0
sqlalchemy>=2.0.0
pyarrow>=14.0.0


# AWS S3

boto3>=1.34.0
botocore>=1.34.0


# Redis & Caching

redis>=5.0.0
hiredis>=2.3.0


# HTTP Client

httpx>=0.26.0
requests>=2.31.0
aiohttp>=3.9.0


# Environment & Configuration

python-dotenv>=1.0.0


# Security & Cryptography

urllib3>=1.26.0,<2.0.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4


# Data Processing

pandas>=2.1.0
numpy>=1.26.0


# Job Scraping (Signals Pipeline)

python-jobspy>=1.1.0


# Testing

pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
fakeredis>=2.21.0
moto>=5.0.0

# HTTP requests for downloading exhibits
requests>=2.31.0

# Optional: Convert HTML filings to PDF (requires system dependencies)
weasyprint>=60.0
